% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agreements.R
\name{agree}
\alias{agree}
\title{Agreements}
\usage{
agree(df1, df2, df3, name, algorithm)
}
\arguments{
\item{df1}{First dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{df2}{Second dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{df3}{Third dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{name}{Name of the algorithm for which your agreements will be shown}

\item{algorithm}{Algorithm from which agreements will be displayed. A 1 shall show the agreements of the first algorithm, a 2 of the
second algorithm and 3 of the third algorithm.}
}
\description{
Function to generate an agreement matrix from the metrics of any three algorithms.
}
\details{
The agreements between the Accuracy, Kappa, F1, AUC, LiftAUC and BA measures for a particular algorithm shall be represented
in a triangular matrix form where:

- The cells that are the diagonal will symbolise the number of times an algorithm has given as best algorithm the
particular algorithm. For example, the position [Accuracy, Accuracy] will represent the number of times that Accuracy has
given the algorithm as the best algorithm.

- The cells that are not the diagonal will symbolise the number of agreements that exist for that pair of metrics for the
particular algorithm. For example, the position [Accuracy, Kappa] will represent the number of agreements that exist between
these characteristics for the algorithm.
}
\examples{

AgreeDisagreeR::agree("IPIP_unbalancing.rds", "SMOTE_unbalancing.rds", "ROSE_unbalancing.rds", "IPIP", 1)
AgreeDisagreeR::agree("IPIP_unbalancing.rds", "SMOTE_unbalancing.rds", "ROSE_unbalancing.rds", "SMOTE", 2)
AgreeDisagreeR::agree("IPIP_unbalancing.rds", "SMOTE_unbalancing.rds", "ROSE_unbalancing.rds", "ROSE", 3)
AgreeDisagreeR::agree("IPIP_unbalancing_complete.rds", "SMOTE_unbalancing_complete.rds", "ROSE_unbalancing_complete.rds", "IPIP", 1)
AgreeDisagreeR::agree("IPIP_unbalancing_complete.rds", "SMOTE_unbalancing_complete.rds", "ROSE_unbalancing_complete.rds", "SMOTE", 2)
AgreeDisagreeR::agree("IPIP_unbalancing_complete.rds", "SMOTE_unbalancing_complete.rds", "ROSE_unbalancing_complete.rds", "ROSE", 3)
}
