% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgreementsDisagreements.R
\name{agree_disagree}
\alias{agree_disagree}
\title{Agreements and Disagreements}
\usage{
agree_disagree(df1, df2, df3, name1, name2, name3)
}
\arguments{
\item{df1}{First dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{df2}{Second dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{df3}{Third dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{name1}{Name of the first algorithm}

\item{name2}{Name of the second algorithm}

\item{name3}{Name of the third algorithm}
}
\description{
Function to generate the agreement and disagreement matrices from the metrics of three different algorithms.
}
\details{
For the explanation of each type of matrix go to the documentation of the functions "agreements" and "disagreements".
}
\examples{

AgreeDisagreeR::agree_disagree("IPIP_unbalancing.rds", "SMOTE_unbalancing.rds", "ROSE_unbalancing.rds", "IPIP", "SMOTE", "ROSE")
AgreeDisagreeR::agree_disagree("IPIP_unbalancing_complete.rds", "SMOTE_unbalancing_complete.rds", "ROSE_unbalancing_complete.rds", "IPIP", "SMOTE", "ROSE")
}
