% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disagreements.R
\name{disagree}
\alias{disagree}
\title{Disagreements}
\usage{
disagree(df1, df2, df3)
}
\arguments{
\item{df1}{First dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{df2}{Second dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{df3}{Third dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}
}
\description{
Function to generate a disagreement matrix from the metrics of any three algorithms.
}
\details{
The disagreements between the Accuracy, Kappa, F1, AUC, LiftAUC and BA metrics between the algorithms shall be represented
in the form of a triangular matrix where each position shall symbolise the number of disagreements for that pair of metrics.
triangular matrix where each position will symbolise the number of disagreements that exist for that pair of metrics for the
three algorithms.

For example, the cell [Accuracy, Kappa] represents the number of times that both metrics did not give the same algorithm as
the best. algorithm the same algorithm.
}
\examples{

AgreeDisagreeR::disagree("IPIP_unbalancing.rds", "SMOTE_unbalancing.rds", "ROSE_unbalancing.rds")
AgreeDisagreeR::disagree("IPIP_unbalancing_complete.rds", "SMOTE_unbalancing_complete.rds", "ROSE_unbalancing_complete.rds")
}
