% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegratedAgreementsDisagreements.R
\name{integrated_agree_disagree}
\alias{integrated_agree_disagree}
\title{Integrated agreements and disagreements}
\usage{
integrated_agree_disagree(df1, df2, df3, name1, name2, name3)
}
\arguments{
\item{df1}{First dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{df2}{Second dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{df3}{Third dataframe that will contain the metrics in lists of experiment lists where for each experiment 5
cross-validation experiments will have been done. The dataframe shall have the value of the metrics in the following order:
Accuracy, Kappa, F1, AUC, LiftAUC and BA.}

\item{name1}{Name of the first algorithm}

\item{name2}{Name of the second algorithm}

\item{name3}{Name of the third algorithm}
}
\description{
Function to generate a matrix of integrated agreements and disagreements from the metrics of any three algorithms.
}
\details{
The integrated agreements and disagreements between the Accuracy, Kappa, F1, AUC, LiftAUC and BA measures between the
algorithms will be represented in the form of a triangular matrix where algorithms in the form of a triangular matrix where:

- In the diagonal, the number of victories that each metric has had for each algorithm is symbolised by a piechart. This
piechart by its definition is always complete.  For example, the box [Accuracy, Accuracy] will represent the number of times
Accuracy has preferred each algorithm.

- In the cells that are not the diagonal, the number of agreements that have existed between all the metric pairs shall be
symbolised with a piechart. the pairs of metrics. For example, the [Accuracy, Kappa] cell shall represent the number of times
 those metrics have agreed in each coloured algorithm and, in white, the number of disagreements of both metrics.
}
\examples{

AgreeDisagreeR::integrated_agree_disagree("IPIP_unbalancing.rds", "SMOTE_unbalancing.rds", "ROSE_unbalancing.rds", "IPIP", "SMOTE", "ROSE")
AgreeDisagreeR::integrated_agree_disagree("IPIP_unbalancing_complete.rds", "SMOTE_unbalancing_complete.rds", "ROSE_unbalancing_complete.rds", "IPIP", "SMOTE", "ROSE")
}
